use uuid instead of email for export
