+++ scripts/export-messages
@@ -1,0 +1,20 @@
+#!/bin/bash
+
+set -eu
+
+SCRIPT_DIR=$(cd "$(dirname "$0")" && pwd)
+. "$SCRIPT_DIR/rpc"
+
+if [ $# -lt 1 ]; then
+  echo "Usage: $0 <target-uuid>" >&2
+  exit 1
+fi
+
+target_id="$1"
+
+rpc <<END
+case Proca.Action.MessageExport.export("$target_id") do
+  {:ok, csv} -> IO.puts(csv)
+  {:error, reason} -> IO.puts(:stderr, reason); System.halt(1)
+end
+END
